Repaired Cache Invalidate issue: Modified the src
Issue: The demo was not working properly on the board for the functionalities that included HDMI RX.
Switching between functions using UART terminal was not working and the screen that received data from the board was displaying what the board was receiving from the PC (working like a second screen).
Solution: Removed the line 210 in video_demo.c that was calling the
function Xil_DCacheInvalidateRange(...). See the documentation for the
function here https://docs.xilinx.com/r/2021.2-English/oslib_rm/Xil_DCacheInvalidateRange?tocId=iIPOwKQwYZFmatzmkU5Itw